+++ ios/Classes/SensitiveClipboardPlugin.h
@@ -1,0 +1,4 @@
+#import <Flutter/Flutter.h>
+
+@interface SensitiveClipboardPlugin : NSObject<FlutterPlugin>
+@end