+++ ios/Classes/SwiftSensitiveClipboardPlugin.swift
@@ -1,0 +1,14 @@
+import Flutter
+import UIKit
+
+public class SwiftSensitiveClipboardPlugin: NSObject, FlutterPlugin {
+  public static func register(with registrar: FlutterPluginRegistrar) {
+    let channel = FlutterMethodChannel(name: "sensitive_clipboard", binaryMessenger: registrar.messenger())
+    let instance = SwiftSensitiveClipboardPlugin()
+    registrar.addMethodCallDelegate(instance, channel: channel)
+  }
+
+  public func handle(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
+    result("iOS " + UIDevice.current.systemVersion)
+  }
+}